type warning fix

+++ cuBQL/traversal/fixedBoxQuery.h
@@ -98,8 +98,8 @@ namespace cuBQL {
             // out of down-travesal and let leaf code pop in.
             break;
 
-          uint32_t n0Idx = node.offset+0;
-          uint32_t n1Idx = node.offset+1;
+          uint32_t n0Idx = (uint32_t)node.offset+0;
+          uint32_t n1Idx = (uint32_t)node.offset+1;
           bvh3f::node_t n0 = bvh.nodes[n0Idx];
           bvh3f::node_t n1 = bvh.nodes[n1Idx];
           bool o0 = queryBox.overlaps(n0.bounds);
